Cleanup code layout
