+++ inst/editor/src/Shiny-Ui-Elements/ShinyCheckboxGroupInput/index.tsx
@@ -1,6 +1,6 @@
 import icon from "../../assets/icons/shinyCheckgroup.png";
 import type { CSSMeasure } from "../../components/Inputs/CSSUnitInput/CSSMeasure";
-import { namedListToItemTypeArray } from "../../components/Inputs/ListInput/useListState";
+import { namedListToItemTypeArray } from "../../components/Inputs/ListInput/namedListUtils";
 import type { NamedList } from "../../ui-node-definitions/inputFieldTypes";
 import { input_checkbox_group } from "../../ui-node-definitions/Shiny/input_checkbox_group";
 import { addEditorInfoToUiNode } from "../utils/add_editor_info_to_ui_node";
