+++ inst/editor/src/components/Inputs/ListInput/useListState.tsx
@@ -5,85 +5,15 @@ import type {
   NamedList,
 } from "../../../ui-node-definitions/inputFieldTypes";
 
-type ItemArray = {
-  id: number;
-  key: string;
-  value: string;
-}[];
+import {
+  getKeyValueMismatches,
+  isValueOnlyMode,
+  namedListToItemTypeArray,
+  swapKeysForObject,
+} from "./namedListUtils";
 
+type ItemArray = ReturnType<typeof namedListToItemTypeArray>;
 type ListMode = "key-value" | "value-only";
-
-/**
- * Go from the internal state representation of an array of items to the
- * external state, which either is a key-value pair as an object or a value-only
- * array
- * @param array Internal representation
- * @param valueOnlyMode Whether or not we're in the value-only mode.
- * @returns An array of strings if in value-only mode, or an object if not
- */
-function itemTypeArrayToNamedList(
-  array: ItemArray,
-  listMode: ListMode
-): NamedList {
-  if (listMode === "value-only") {
-    return array.map((item) => item.value);
-  }
-
-  return array.reduce((list, { key, value }) => {
-    list[key] = value;
-    return list;
-  }, {} as Record<string, string>);
-}
-
-/**
- * Convert a NamedList to an array of items for iterating over.
- * @param list NamedList to convert. Could be a key-value object or an array of strings
- * @returns An array of items with an id, key, and value
- */
-export function namedListToItemTypeArray(list: NamedList): ItemArray {
-  if (Array.isArray(list)) {
-    return list.map((value, i) => ({ id: i, key: value, value }));
-  }
-
-  return Object.keys(list).map((key, i) => ({
-    id: i,
-    key,
-    value: list[key],
-  }));
-}
-
-function isValueOnlyMode(value: NamedList): value is string[] {
-  return Array.isArray(value);
-}
-
-/**
- * Update the key of an object while making sure the position of the item in the
- * object is preserved
- * @param obj object keyed by a string
- * @param oldKey Previous key value
- * @param newKey Key value to update to
- * @returns A new object with the key updated to new value
- */
-function swapKeysForObject<ValueType>(
-  obj: Record<string, ValueType>,
-  oldKey: string,
-  newKey: string
-): Record<string, ValueType> {
-  // Convert object to array of key-value pairs
-  const entries = Object.entries(obj);
-
-  // Find the index of the item with the old key
-  const index = entries.findIndex(([key]) => key === oldKey);
-
-  // Update the key of the item at the index
-  entries[index][0] = newKey;
-
-  // Repack the array of key-value pairs into an object
-  return entries.reduce((obj, [key, value]) => {
-    obj[key] = value;
-    return obj;
-  }, {} as Record<string, ValueType>);
-}
 
 export function useListState({
   value,
@@ -262,16 +192,24 @@ export function useListState({
   };
 }
 
-function getKeyValueMismatches(x: Record<string, unknown>): ItemArray | null {
-  const mismatches = Object.entries(x)
-    .filter(([key, value]) => typeof value === "string" && key !== value)
-    .map(([key, value], i) => ({ key, value, id: i }));
-
-  if (mismatches.length === 0) {
-    return null;
+/**
+ * Go from the internal state representation of an array of items to the
+ * external state, which either is a key-value pair as an object or a value-only
+ * array
+ * @param array Internal representation
+ * @param valueOnlyMode Whether or not we're in the value-only mode.
+ * @returns An array of strings if in value-only mode, or an object if not
+ */
+function itemTypeArrayToNamedList(
+  array: ItemArray,
+  listMode: ListMode
+): NamedList {
+  if (listMode === "value-only") {
+    return array.map((item) => item.value);
   }
 
-  // Forcing this type is okay because it already has been type checked because we said
-  // key
-  return mismatches as ItemArray;
+  return array.reduce((list, { key, value }) => {
+    list[key] = value;
+    return list;
+  }, {} as Record<string, string>);
 }
